feat: SessionChat.tsx & getChatAccess.ts Added SessionChat.tsx component for more readability/reusability. Added getChatAccess.ts to extract the logic for chat access.
