+++ components/storyblok/StoryblokSessionIbaPage.tsx
@@ -1,6 +1,5 @@
-import ChatBubbleOutlineIcon from '@mui/icons-material/ChatBubbleOutline';
 import StarBorderIcon from '@mui/icons-material/StarBorder';
-import { Box, Container, Typography } from '@mui/material';
+import { Box, Container } from '@mui/material';
 import { ISbRichtext, ISbStoryData, storyblokEditable } from '@storyblok/react';
 import { useTranslations } from 'next-intl';
 import Head from 'next/head';
@@ -9,18 +8,18 @@ import { render } from 'storyblok-rich-text-react-renderer';
 import { PROGRESS_STATUS } from '../../constants/enums';
 import { useTypedSelector } from '../../hooks/store';
 import { columnStyle } from '../../styles/common';
-import hasAccessToPage from '../../utils/hasAccessToPage';
 import { getEventUserData } from '../../utils/logEvent';
 import { RichTextOptions } from '../../utils/richText';
 import SessionContentCard from '../cards/SessionContentCard';
 import { Dots } from '../common/Dots';
-import CrispButton from '../crisp/CrispButton';
 import MultipleBonusContent, { BonusContent } from '../session/MultipleBonusContent';
 import { SessionCompleteButton } from '../session/SessionCompleteButton';
-import Video from '../video/Video';
 import { getSessionCompletion } from '../../utils/getSessionCompletion';
 import { SessionHeader } from '../session/SessionHeader';
 import { SessionVideo } from '../session/SessionVideo';
+import { getChatAccess } from '../../utils/getChatAccess';
+import { SessionChat } from '../session/SessionChat';
+import hasAccessToPage from '../../utils/hasAccessToPage';
 
 const containerStyle = {
   backgroundColor: 'secondary.light',
@@ -31,14 +30,6 @@ const cardColumnStyle = {
   alignItems: 'center',
   gap: { xs: 2, md: 3 },
 } as const;
-
-const crispButtonContainerStyle = {
-  paddingTop: 4,
-  paddingBottom: 1,
-  display: 'flex',
-} as const;
-
-const chatDetailIntroStyle = { marginTop: 3, marginBottom: 1.5 } as const;
 
 export interface StoryblokSessionIbaPageProps {
   storyId: number;
@@ -78,31 +69,22 @@ const StoryblokSessionIbaPage = (props: StoryblokSessionIbaPageProps) => {
   const t = useTranslations('Courses');
 
   const userCreatedAt = useTypedSelector((state) => state.user.createdAt);
-  const userEmail = useTypedSelector((state) => state.user.email);
   const isLoggedIn = useTypedSelector((state) => Boolean(state.user.id));
   const partnerAccesses = useTypedSelector((state) => state.partnerAccesses);
   const partnerAdmin = useTypedSelector((state) => state.partnerAdmin);
   const courses = useTypedSelector((state) => state.courses);
 
   const [incorrectAccess, setIncorrectAccess] = useState<boolean>(true);
-  const [liveChatAccess, setLiveChatAccess] = useState<boolean>(false);
   const [sessionProgress, setSessionProgress] = useState<PROGRESS_STATUS>(
     PROGRESS_STATUS.NOT_STARTED,
   );
   const [weekString, setWeekString] = useState<string>('');
 
-  // TODO refactor chat access logic
   useEffect(() => {
     const coursePartners = course.content.included_for_partners;
     setIncorrectAccess(
       !hasAccessToPage(isLoggedIn, false, coursePartners, partnerAccesses, partnerAdmin),
     );
-
-    const liveAccess = partnerAccesses.find(
-      (partnerAccess) => partnerAccess.featureLiveChat === true,
-    );
-
-    if (liveAccess) setLiveChatAccess(true);
   }, [partnerAccesses, course.content.included_for_partners, partnerAdmin]);
 
   useEffect(() => {
@@ -152,18 +134,14 @@ const StoryblokSessionIbaPage = (props: StoryblokSessionIbaPageProps) => {
           />
           <Container sx={containerStyle}>
             <Box sx={cardColumnStyle}>
-              {video && (
-                <>
-                  <SessionVideo
-                    eventData={eventData}
-                    name={name}
-                    video={video}
-                    storyId={storyId}
-                    sessionProgress={sessionProgress}
-                    video_transcript={video_transcript}
-                  />
-                </>
-              )}
+              <SessionVideo
+                eventData={eventData}
+                name={name}
+                video={video}
+                storyId={storyId}
+                sessionProgress={sessionProgress}
+                video_transcript={video_transcript}
+              />
               {activity.content &&
                 (activity.content?.length > 1 || activity.content[0].content) && (
                   <>
@@ -180,46 +158,11 @@ const StoryblokSessionIbaPage = (props: StoryblokSessionIbaPageProps) => {
                   </>
                 )}
               {bonus && <MultipleBonusContent bonus={bonus} eventData={eventData} />}
-              {liveChatAccess && (
-                <>
-                  <Dots />
-                  <SessionContentCard
-                    title={t('sessionDetail.chat.title')}
-                    titleIcon={ChatBubbleOutlineIcon}
-                    titleIconSize={24}
-                    eventPrefix="SESSION_CHAT"
-                    eventData={eventData}
-                  >
-                    <Typography paragraph>{t('sessionDetail.chat.description')}</Typography>
-                    <Typography paragraph>{t('sessionDetail.chat.videoIntro')}</Typography>
-                    <Video
-                      eventPrefix="SESSION_CHAT_VIDEO"
-                      eventData={eventData}
-                      url={t('sessionDetail.chat.videoLink')}
-                      containerStyles={{ mx: 'auto', my: 2 }}
-                    ></Video>
-                    <Box sx={chatDetailIntroStyle}>
-                      <Typography>{t('sessionDetail.chat.detailIntro')}</Typography>
-                    </Box>
-                    <Box>
-                      <ul>
-                        <li>{t('sessionDetail.chat.detailPrivacy')}</li>
-                        <li>{t('sessionDetail.chat.detailTimezone')}</li>
-                        <li>{t('sessionDetail.chat.detailLanguage')}</li>
-                        <li>{t('sessionDetail.chat.detailLegal')}</li>
-                        <li>{t('sessionDetail.chat.detailImmediateHelp')}</li>
-                      </ul>
-                    </Box>
-                    <Box sx={crispButtonContainerStyle}>
-                      <CrispButton
-                        email={userEmail}
-                        eventData={eventData}
-                        buttonText={t('sessionDetail.chat.startButton')}
-                      />
-                    </Box>
-                  </SessionContentCard>
-                </>
-              )}
+              <SessionChat
+                eventData={eventData}
+                partnerAccesses={partnerAccesses}
+                partnerAdmin={partnerAdmin}
+              />
               {sessionProgress !== PROGRESS_STATUS.COMPLETED && (
                 <SessionCompleteButton storyId={storyId} eventData={eventData} />
               )}
